Avoid showing duplicate error message
The user was getting an error message several times if the result was negative. This happened each time the subtract button was being clicked.

A class check solves the bug. A class has been added in the p element so that the program can detect it. If there's already a message, the program won't print it out again.